:camera_flash: Change RIT logo only with white color (closes #4)
This commit changes the logo the RIT Logo White variant provided from
the brand portal:

    https://www.rit.edu/marketing/brandportal/toolkits/logos

I tried other variants and this is the one that looked the best with
minimal changes. I did not want to make changes to the upstream theme
because I would need to fork the theme and update the git submodules for
this one change, which is not something I am prepared to do at this
phase of this project.

This can be revisited later once the future of this site is clear. For
now, it is only a proof-of-concept and might not even be used. So, I am
not interested in spending too much time on trying to get this one
aspect correct at the current moment.

Closes #4.

Signed-off-by: Justin W. Flory <git@jwf.io>